+++ packages/abi/lib/coders/array.js
@@ -210,6 +210,17 @@ var ArrayCoder = /** @class */ (function (_super) {
         var count = this.length;
         if (count === -1) {
             count = reader.readValue().toNumber();
+            // Check that there is *roughly* enough data to ensure
+            // stray random data is not being read as a length. Each
+            // slot requires at least 32 bytes for their value (or 32
+            // bytes as a link to the data). This could use a much
+            // tighter bound, but we are erroring on the side of safety.
+            if (count * 32 > reader._data.length) {
+                logger.throwError("insufficient data length", logger_1.Logger.errors.BUFFER_OVERRUN, {
+                    length: reader._data.length,
+                    count: count
+                });
+            }
         }
         var coders = [];
         for (var i = 0; i < count; i++) {