+++ packages/bignumber/lib.esm/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "bignumber/5.1.0";
+export declare const version = "bignumber/5.1.1";
 //# sourceMappingURL=_version.d.ts.map