+++ packages/providers/lib/_version.d.ts
@@ -1,2 +1,2 @@
-export declare const version = "providers/5.1.1";
+export declare const version = "providers/5.1.2";
 //# sourceMappingURL=_version.d.ts.map