+++ packages/providers/lib/json-rpc-provider.js
@@ -354,7 +354,6 @@ var JsonRpcProvider = /** @class */ (function (_super) {
             });
         }
         _this = _super.call(this, networkOrReady) || this;
-        _this._eventLoopCache = {};
         // Default URL
         if (!url) {
             url = properties_1.getStatic(_this.constructor, "defaultUrl")();
@@ -370,19 +369,29 @@ var JsonRpcProvider = /** @class */ (function (_super) {
         _this._nextId = 42;
         return _this;
     }
+    Object.defineProperty(JsonRpcProvider.prototype, "_cache", {
+        get: function () {
+            if (this._eventLoopCache == null) {
+                this._eventLoopCache = {};
+            }
+            return this._eventLoopCache;
+        },
+        enumerable: false,
+        configurable: true
+    });
     JsonRpcProvider.defaultUrl = function () {
         return "http:/\/localhost:8545";
     };
     JsonRpcProvider.prototype.detectNetwork = function () {
         var _this = this;
-        if (!this._eventLoopCache["detectNetwork"]) {
-            this._eventLoopCache["detectNetwork"] = this._uncachedDetectNetwork();
+        if (!this._cache["detectNetwork"]) {
+            this._cache["detectNetwork"] = this._uncachedDetectNetwork();
             // Clear this cache at the beginning of the next event loop
             setTimeout(function () {
-                _this._eventLoopCache["detectNetwork"] = null;
+                _this._cache["detectNetwork"] = null;
             }, 0);
         }
-        return this._eventLoopCache["detectNetwork"];
+        return this._cache["detectNetwork"];
     };
     JsonRpcProvider.prototype._uncachedDetectNetwork = function () {
         return __awaiter(this, void 0, void 0, function () {
@@ -462,8 +471,8 @@ var JsonRpcProvider = /** @class */ (function (_super) {
         // We can expand this in the future to any call, but for now these
         // are the biggest wins and do not require any serializing parameters.
         var cache = (["eth_chainId", "eth_blockNumber"].indexOf(method) >= 0);
-        if (cache && this._eventLoopCache[method]) {
-            return this._eventLoopCache[method];
+        if (cache && this._cache[method]) {
+            return this._cache[method];
         }
         var result = web_1.fetchJson(this.connection, JSON.stringify(request), getResult).then(function (result) {
             _this.emit("debug", {
@@ -484,9 +493,9 @@ var JsonRpcProvider = /** @class */ (function (_super) {
         });
         // Cache the fetch, but clear it on the next event loop
         if (cache) {
-            this._eventLoopCache[method] = result;
+            this._cache[method] = result;
             setTimeout(function () {
-                _this._eventLoopCache[method] = null;
+                _this._cache[method] = null;
             }, 0);
         }
         return result;